backend/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "python-dotenv>=1.1.0,<2.0",
  "fastmcp>=2.12.0,<3.0",
  "SQLAlchemy>=2.0.0,<3.0",
  "stytch>=13.20.0,<14.0",
  "uvicorn>=0.35.0,<1.0",
  "python-jose>=3.5.0,<4.0",
]
